// frontend/src/shared/lib/api/model/datasetPreviewResponse.ts
/**
 * Generated by orval. Do not edit manually.
 * daikon-studio
 */
import type { DatasetPreviewResponseStatus } from './datasetPreviewResponseStatus';
import type { DatasetPreviewResponseDatasetId } from './datasetPreviewResponseDatasetId';
import type { DatasetPreviewResponseError } from './datasetPreviewResponseError';
import type { DatasetPreviewResponsePreparation } from './datasetPreviewResponsePreparation';

export interface DatasetPreviewResponse {
  id: string;
  name: string;
  status: DatasetPreviewResponseStatus;
  stage: string;
  done: number;
  total: number;
  dataset_id: DatasetPreviewResponseDatasetId;
  error: DatasetPreviewResponseError;
  created_at: string;
  updated_at: string;
  preparation: DatasetPreviewResponsePreparation;
}